+++ test/joken_token_test.exs
@@ -348,7 +348,7 @@ defmodule Joken.Token.Test do
       end
 
       def claim(:exp, _payload) do
-        Joken.Config.get_current_time() + 300
+        Joken.Helpers.get_current_time() + 300
       end
 
       def claim(_, _) do
@@ -356,7 +356,7 @@ defmodule Joken.Token.Test do
       end
 
       def validate_claim(:exp, payload, _) do
-        Joken.Config.validate_time_claim(payload, :exp, "Token expired", fn(expires_at, now) -> expires_at > now end)
+        Joken.Helpers.validate_time_claim(payload, :exp, "Token expired", fn(expires_at, now) -> expires_at > now end)
       end
 
       def validate_claim(_, _, _) do
@@ -384,7 +384,7 @@ defmodule Joken.Token.Test do
       end
 
       def claim(:exp, _payload) do
-        Joken.Config.get_current_time() - 300
+        Joken.Helpers.get_current_time() - 300
       end
 
       def claim(_, _) do
@@ -392,7 +392,7 @@ defmodule Joken.Token.Test do
       end
 
       def validate_claim(:exp, payload, _) do
-        Joken.Config.validate_time_claim(payload, :exp, "Token expired", fn(expires_at, now) -> expires_at > now end)
+        Joken.Helpers.validate_time_claim(payload, :exp, "Token expired", fn(expires_at, now) -> expires_at > now end)
       end
 
       def validate_claim(_, _, _) do
@@ -431,7 +431,7 @@ defmodule Joken.Token.Test do
       end
 
       def claim(:iat, _payload) do
-        Joken.Config.get_current_time() - 300
+        Joken.Helpers.get_current_time() - 300
       end
 
       def claim(_, _) do
@@ -439,7 +439,7 @@ defmodule Joken.Token.Test do
       end
 
       def validate_claim(:iat, payload, _) do
-        Joken.Config.validate_time_claim(payload, :iat, "Token not valid yet", fn(not_before, now) -> not_before < now end) 
+        Joken.Helpers.validate_time_claim(payload, :iat, "Token not valid yet", fn(not_before, now) -> not_before < now end) 
       end
 
       def validate_claim(_, _, _) do
@@ -467,7 +467,7 @@ defmodule Joken.Token.Test do
       end
 
       def claim(:iat, _payload) do
-        Joken.Config.get_current_time() + 300
+        Joken.Helpers.get_current_time() + 300
       end
 
       def claim(_, _) do
@@ -475,7 +475,7 @@ defmodule Joken.Token.Test do
       end
 
       def validate_claim(:iat, payload, _) do
-        Joken.Config.validate_time_claim(payload, :iat, "Token not valid yet", fn(not_before, now) -> not_before < now end) 
+        Joken.Helpers.validate_time_claim(payload, :iat, "Token not valid yet", fn(not_before, now) -> not_before < now end) 
       end
 
       def validate_claim(_, _, _) do